+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## 2020-07-30
 
+### Added
+
+- Elastic APM reporting to the proxy.
+
 ### Changed
 
 - Bump mobius3 to version that better supports EFS